% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2dfuncs.R
\name{autocorr}
\alias{autocorr}
\title{compute autocorrelation matrix of img using fast fourier transforms}
\usage{
autocorr(x)
}
\arguments{
\item{x}{- n x n raw image matrix}
}
\value{
- n x n autocorrelation matrix
}
\description{
compute autocorrelation matrix of img using fast fourier transforms
}
