% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coarsen}
\alias{coarsen}
\title{coarsen image}
\usage{
coarsen(myfield, nk)
}
\arguments{
\item{myfield}{, N x 3 column matrix [X,Y,data]}

\item{nk}{, coarsen over nk x nk pixels}
}
\value{
coarsened (N/nk x N/nk) x 3 matrix
}
\description{
coarsen image
}
