% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_specd}
\alias{cov_specd}
\title{weighted frequency of a vector}
\usage{
cov_specd(rvec, rprof, r)
}
\arguments{
\item{rvec}{unique distance vector}

\item{rprof}{radial profile}

\item{r}{distance matrix}
}
\value{
empirical spectral covariance
}
\description{
weighted frequency of a vector
}
