% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_euclid}
\alias{dist_euclid}
\title{computes distance matrix, based on euclidean distance}
\usage{
dist_euclid(locs)
}
\arguments{
\item{locs}{n x p matrix of location coordinates in R^p}
}
\value{
n x n distance matrix
}
\description{
computes distance matrix, based on euclidean distance
}
