% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_transform}
\alias{dist_transform}
\title{distance to nearest pixel containing material}
\usage{
dist_transform(myfield, matl)
}
\arguments{
\item{myfield}{, N x 3 column matrix [X,Y,data]}

\item{matl}{, indiate if 0 or 1 is material}
}
\value{
N x 1 vector of distance transforms by pixel
}
\description{
distance to nearest pixel containing material
}
