% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emp_vario}
\alias{emp_vario}
\title{computes empircal variogram given covariance matrix}
\usage{
emp_vario(cov)
}
\arguments{
\item{cov}{n x n matrix}
}
\value{
empirical variogram n x n matrix
}
\description{
computes empircal variogram given covariance matrix
}
