% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dimex.R
\name{fitgen_sim}
\alias{fitgen_sim}
\title{Fit and Generate random image with nonstationary correlations structure}
\usage{
fitgen_sim(locs_obs, locs_new, datamat, per = 0.05, lambdas, qthresh,
  seed)
}
\arguments{
\item{locs_obs}{- coordinates of observed data, n x 2 matrix}

\item{locs_new}{- new indices at which to generate new image, N x 2 matrix}

\item{datamat}{- N x m matrix of m replicates of size n at observed locations}

\item{per}{- proportion of points n to uniformly sample across image, default p=0.05}

\item{lambdas}{- list specify tuning parameters lambda1 and lambda2}

\item{qthresh}{- quantile to threshold at}

\item{seed}{- random seed for generation, default NA}
}
\value{
- prints GRF before and after thresholding plus estimated Z (N x 5 matrix)
}
\description{
generates random image of size N x N using parameters fit to observed data of size N x N
}
