% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_exp.R
\name{fitgen_sim_stat}
\alias{fitgen_sim_stat}
\title{Fit model and generate random image with stationary exponential correlation structure}
\usage{
fitgen_sim_stat(datamat, per = 0.05, locs_obs, locs_new, qthresh,
  seed = NA)
}
\arguments{
\item{datamat}{- N x m matrix of m replicates of size n at observed locations}

\item{per}{- proportion of points n to uniformly sample across image, default p=0.05}

\item{locs_obs}{- coordinates of observed data, N x 2 matrix}

\item{locs_new}{- new indices at which to generate new image, N x 2 matrix}

\item{qthresh}{- quantile to threshold at}

\item{seed}{- random seed for generation, default NA}
}
\value{
- prints GRF before and after thresholding (N x 4 matrix)
}
\description{
generates random image of size N x N using parameters fit to observed data of size N x N
}
