% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitkrig}
\alias{fitkrig}
\title{Simple Kriging}
\usage{
fitkrig(data, locs_new, locs_obs, sig_obs, params)
}
\arguments{
\item{data}{n x 1 vector of observed values at locs_obs}

\item{locs_new}{k x p matrix of k new locations}

\item{locs_obs}{n x p matrix of observed locations}

\item{sig_obs}{n x n covariance matrix of observed data}

\item{params}{vector of covaraince parameter values, c(sigma2, phi)}
}
\value{
empirical variogram, nx n matrix
}
\description{
Simple Kriging
}
