% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{freqweight}
\alias{freqweight}
\title{weighted frequency of a vector}
\usage{
freqweight(x, w)
}
\arguments{
\item{x}{vector of values}

\item{w}{vector of weights}
}
\value{
weighted contingency table
}
\description{
weighted frequency of a vector
}
