% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dimex.R
\name{gen_dimex}
\alias{gen_dimex}
\title{Generate random image with nonstationary correlations structure}
\usage{
gen_dimex(locs_obs, locs_new, lambda2, theta, qthresh, seed = NA,
  par = FALSE)
}
\arguments{
\item{locs_obs}{- coordinates of observed data, n x 2 matrix}

\item{locs_new}{- new indices at which to generate new image, N x 2 matrix}

\item{lambda2}{- tuning parameter lambda2.}

\item{theta}{- [sigma2, phi, Z] params from fit_dimex()}

\item{qthresh}{- quantile to threshold at}

\item{seed}{- random seed for generation, default NA}

\item{par}{- true/false if to parallelize, default=FALSE. should not be TRUE if generating multiple GRFs in parallel already. should=TRUE if generating one image at a time.}
}
\value{
- prints GRF before and after thresholding plus estimated Z (N x 5 matrix)
}
\description{
generates random image of size N x N using parameters fit to observed data of size N x N
}
