% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_datamat}
\alias{make_datamat}
\title{grids entire image into m snips of size snipsize}
\usage{
make_datamat(myfield, snipsize, nsmooth)
}
\arguments{
\item{myfield}{, N x 3 column matrix [X,Y,data]}

\item{snipsize}{, size of snipped image (n)}
}
\value{
n^2 x m matrix of m replicates
}
\description{
grids entire image into m snips of size snipsize
}
