% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_z_rcpp}
\alias{optim_z_rcpp}
\title{optim function to estimate extra dimension Z}
\usage{
optim_z_rcpp(init_val, locs_obs, emp_vario, lambda)
}
\arguments{
\item{init_val}{n+2 vector with initial values for [exp(sigma2),exp(phi),Z]}

\item{locs_obs}{observed coordinates n x 2 matrix}

\item{emp_vario}{n x n empirical variogram matrix}

\item{lambda}{tuning parameter lambda1}
}
\value{
n+2 vector with estimated exponential covariance parameters and Z
}
\description{
optim function to estimate extra dimension Z
}
