% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prof_nll}
\alias{prof_nll}
\title{profile negative log-likelihood}
\usage{
prof_nll(phi, datamat, ds, prange)
}
\arguments{
\item{datamat}{n x m matrix}

\item{ds}{n x n matrix}

\item{prange}{practical range, if unsure set to max(ds)}
}
\value{
estimated phi
}
\description{
profile negative log-likelihood
}
