% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2dfuncs.R
\name{radial_profile}
\alias{radial_profile}
\title{compute radial profile of raw image}
\usage{
radial_profile(locs, x, sigd = 0)
}
\arguments{
\item{locs}{x 2 matrix of locations [x,y] #column name matters}

\item{x}{x 1 vector of observed value}
}
\value{
- list of 2: vector of distances and vector of respective radial profile values
}
\description{
compute radial profile of raw image
}
